refactor *interp2D into generic function
ERA, MERRA, and JRA now share code for interp2d.

Some blocks within interp2d broken out into separate functions
